server/requirements.txt
coreapi==2.3.3
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1